requirements.txt
gitpython
langfuse